task: assign gtid to each task.
We assign gtid (i.e., kmp_base_info) to each task when it is executed.
gtid is basically chosen from a blocked thread in the team.  However, if
no thread is blocked in the team, we reuse the master thread's gtid.  We
will need to check the safety of this approach if it causes a problem.