task: allow using gtid not having a valid ULT.
In the thread info structure (i.e., kmp_info), Argobots ULT may not have
been created or revived because the master thread in the team creates
ULTs at the last moment.  This caused a data race when trying to read
the state of associated ULT in the kmp_info.  This patch fixes the
problem by allowing to use the kmp_info that does not have a valid ULT.
This approach should be working because we just use the kmp_info, not
the ULT associated with it.