kmp_info: only one ULT uses it at a certain time
Since running more than one thread that share the same thread
information (i.e., kmp_info) at the same time is not correct in the
current OpenMP standard, we allow only one ULT, whether it is an OpenMP
thread or task, to use kmp_info at a certain time.  To do this, we
utilize the th_active field in the kmp_info struct to synchronize
parallel accesses to kmp_info.