task: manage tasks in a hierarchical manner
We used to save all tasks in a single queue in kmp_info regardless of
their level.  As a result, it was not easy to implement taskwait
correctly because figuring out exact child tasks from the current task
was almost impossible.  This patch adds a task queue to each taskdata
instead of having a single queue in kmp_info and manages tasks in a
hierarchical manner.  With this hierarchical way, a task can easily wait
for the completion of its child tasks.