[openmp][cmake] passing option argument correctly
From the context, it looks like the test should not be run with `check-all`,
but it does. It turns out option argument resolving to True/False which
could not be passed down as is. There is one such example in
AddLLVM.cmake.

cherry-pick: c2c4f1c1202ab5d66b3cf4b1d9769f3a8572d321
https://github.com/llvm/llvm-project/commit/c2c4f1c1202ab5d66b3cf4b1d9769f3a8572d321