[OpenMP] target_data_begin: fail on device alloc fail
Without this patch, target_data_begin continues after an illegal
mapping or an out-of-memory error on the device.  With this patch, it
terminates the runtime with an error instead.

The new test exercises only illegal mappings.  I didn't think of a
good way to exercise out-of-memory errors from the test suite.

Reviewed By: jdoerfert

Differential Revision: https://reviews.llvm.org/D78170

cherry-pick: 5f6aa9680c19534a382cb652178e0ff6edbe78de
https://github.com/llvm/llvm-project/commit/5f6aa9680c19534a382cb652178e0ff6edbe78de