[OpenMP] Fix a race in task queue reallocation
__kmp_realloc_task_deque implicitly assumes, that the task queue is full
(ntasks == size), therefore tail = size in line 319.
An assertion is added to document this assumption.

The first check for a full queue is before the locking and might not hold
when the lock is taken. So, we need to check again for this condition when
we have the lock.

Reviewed By: AndreyChurbanov

Differential Revision: https://reviews.llvm.org/D80480

cherry-pick: 3895148d7cd8ff76220f8f8209ec06369a8e816f
https://github.com/llvm/llvm-project/commit/3895148d7cd8ff76220f8f8209ec06369a8e816f