[libomptarget][NFC] Replace `%ld` with PRId64 for data of type int64_t.
The standard way of printing `int64_t` data is via the PRId64 macro, `ld`
is for `long int` and int64_t is not guaranteed to be typedef'ed as `long int`
on all platforms. E.g. on Windows we get mismatch warnings.

Differential Revision: https://reviews.llvm.org/D85353

cherry-pick: 40470eb27a5c97b01e89d8825626487b0682abec
https://github.com/llvm/llvm-project/commit/40470eb27a5c97b01e89d8825626487b0682abec