[OpenMP][OMPT] Update OMPT tests for newly added GOMP interface patches
This patch updates the expected results for the GOMP interface patches: D87267, D87269, and D87271.
The taskwait-depend test is changed to really use taskwait-depend and copied to an task_if0-depend test.

To pass the tests, the handling of the return address was fixed.

Differential Revision: https://reviews.llvm.org/D87680

cherry-pick: 6104b30446aa976006fd322af4a57a8f0124f94f
https://github.com/llvm/llvm-project/commit/6104b30446aa976006fd322af4a57a8f0124f94f