[BOLT] better support for in-tree build
The current LLVM project is maintained in a single repository.  Although BOLT
will be provided as a single repository, BOLT should work with the LLVM
repository, too.  This patch updates CMakefiles so that BOLT can be separately
installed as "BOLT" in addition to the default LLVM OpenMP library.  BOLT will
be installed to bolt/ directory by default unless you set
-DLIBOMP_USE_BOLT_DEFAULT=true.

The original LLVM project does not recognize BOLT, so the user needs to modify
its CMakefiles.  This patch will be created in the future.