[OpenMP] Fixed a potential integer overflow
`size_t` has different width on 32- and 64-bit architecture, but the
computation to floor to power of two assumed it is 64-bit, which can cause an
integer overflow. In this patch, architecture detection is added so that the
operation for 64-bit `size_t`. Thank Luke for reporting the issue.

Reviewed By: jdoerfert

Differential Revision: https://reviews.llvm.org/D89878

cherry-pick: 3091ed099f2f6a3d16dbdae7d0406f54dfc3031f
https://github.com/llvm/llvm-project/commit/3091ed099f2f6a3d16dbdae7d0406f54dfc3031f