[OpenMP] Begin Adding OpenMP Tool to Gather OpenMP Information
Summary:
This patch begins to add support for a set of scripts that can be used to get information from OpenMP programs to better describe problems and eventually show the data to the user in formatted output. Right now the only support is forformatting the register and memory usage reports from ptxas and nvlink. This is simply done as a wrapper around clang and clang++.

Reviewers: jdoerfert

DIfferential Revision: https://reviews.llvm.org/D91085

cherry-pick: 292e898c16d120bdb629e0e7ba743c0aa1b9f1d2
https://github.com/llvm/llvm-project/commit/292e898c16d120bdb629e0e7ba743c0aa1b9f1d2