[OpenMP][deviceRTLs] Remove omp_is_initial_device
`omp_is_initial_device` in device code was implemented as a builtin
function in D38968 for a better performance. Therefore there is no chance that
this function will be called to `deviceRTLs`. As we're moving to build `deviceRTLs`
with OpenMP compiler, this function can lead to a compilation error. This patch
just simply removes it.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D95397

cherry-pick: 3333244d77c44e8bb5af57027646596f7714ff62
https://github.com/llvm/llvm-project/commit/3333244d77c44e8bb5af57027646596f7714ff62